lab13/homework/pyproject.toml
[project]
name = "MLOps_course_AGH_lab_13_homework"
version = "1.0.0"
description = ""
requires-python = "==3.11.*"

dependencies = [
    "guardrails-ai>=0.7.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mcp-use>=1.5.2",
    "openai<2", # Guardrails AI limitation
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
linux = [
    "ruff",
    "vllm>=0.13.0",
    "bitsandbytes>=0.49.1",
    "transformers[torch]>=4.51.0",
    "fastmcp>=2.14.3",
]
mac = [
    "fastmcp>=2.14.3",
    "transformers[torch]>=4.51.0",
    "ruff"
]
